Move langgraph data generation to integration tests
