+++ tests/unit/test_data_loader.py
@@ -1,10 +1,10 @@
-import unittest
-import tempfile
 import pathlib
+import tempfile
+import unittest
 
+from flexeval import data_loader
 from flexeval.classes.dataset import Dataset
 from flexeval.classes.eval_runner import EvalRunner
-from flexeval import data_loader
 
 
 class TempPathMixin:
